(#%top t.meth) generates a thunk or a curried procedure
Basically, we want to ensure the first argument to the method is the table, which
we know at the site whene t.meth appears. I think this takes care of methods of
arity 1 (that only take self) and higher arities.

Somehow I doubt this works for methods that also accept keyword arguments. I'll
have to experiment and think about that.